ISA_SuperApp/etl/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "isa-etl"
version = "1.0.0"
description = "ISA_D Enterprise ETL System with Dagster and OpenLineage"
authors = [
    {name = "ISA_D Team", email = "team@isa-d.org"},
]
dependencies = [
    "dagster==1.7.14",
    "dagster-webserver==1.7.14",
    "dagster-postgres==0.23.14",
    "openlineage-integration-common==1.15.0",
    "openlineage-dagster==1.8.0",
    "requests==2.31.0",
    "pandas==2.1.4",
    "sqlalchemy==2.0.23",
    "psycopg2-binary==2.9.9",
    "pydantic==2.5.2",
    "python-dotenv==1.0.0",
    "structlog==23.2.0",
    "prometheus-client==0.19.0",
    "datadog==0.48.0",
    "sentry-sdk==1.38.0",
    "schedule==1.2.1",
    "tenacity==8.2.3",
]
requires-python = ">=3.8"

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-cov==4.1.0",
    "black==23.12.1",
    "isort==5.13.2",
    "flake8==6.1.0",
    "mypy==1.7.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["isa_etl*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true